requirement.txt
# requirements.txt 예시
# Python 3.9+ 권장

#pip install torch==2.0.1 torchvision==0.15.2 torchaudio==2.0.2 --index-url https://download.pytorch.org/whl/cu118


# Transformers & Hugging Face 생태계
transformers==4.30.2
accelerate==0.20.3
peft==0.4.0

# Hydra & Lightning
hydra-core==1.3.2
omegaconf==2.3.0
pytorch-lightning==2.0.6

# OpenCV
opencv-python==4.8.0.74

# 기타
numpy==1.23.5
